+++ stdlib/lib.core.d.ts
@@ -1758,8 +1758,8 @@ interface EventConstructor {
 declare var Event: EventConstructor;
 
 interface EventTarget {
-    addEventListener(type: string, listener: (this: EventTarget, event: Event) => void, options?: boolean | AddEventListenerOptions): void;
-    removeEventListener(type: string, listener: (this: EventTarget, event: Event) => void, options?: boolean | EventListenerOptions): void;
+    addEventListener(type: string, listener: (this: EventTarget, event: Event) => void, options?: boolean | AddEventListenerOptions, ...ignored: any[]): void;
+    removeEventListener(type: string, listener: (this: EventTarget, event: Event) => void, options?: boolean | EventListenerOptions, ...ignored: any[]): void;
     dispatchEvent(event: Event): boolean;
 }
 interface EventTargetConstructor {